# docs/admin/swagger.yaml
basePath: /api/admin
definitions:
  base_info.Success:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      msg:
        example: err msg
        type: string
      reason:
        description: 错误原因，定义为业务判定错误码
        example: success
        type: string
    type: object
  base_info.Swagger400Resp:
    properties:
      code:
        example: 400
        type: integer
      msg:
        example: err msg
        type: string
      reason:
        example: err_reason
        type: string
    type: object
  base_info.Swagger401Resp:
    properties:
      code:
        example: 401
        type: integer
      msg:
        example: err msg
        type: string
      reason:
        example: err_reason
        type: string
    type: object
  base_info.Swagger500Resp:
    properties:
      code:
        example: 500
        type: integer
      msg:
        example: err msg
        type: string
      reason:
        example: err_reason
        type: string
    type: object
...
  commands.CreateRoleCommand:
    properties:
      code:
        description: 角色编码
        type: string
      description:
        description: 描述
        type: string
      localize:
        description: 多语言标识
        type: string
      name:
        description: 角色名称
        type: string
      sequence:
        description: 排序
        type: integer
      type:
        description: 角色类型
        type: integer
    required:
    - code
    - name
    - type
    type: object
...
  commands.UpdateRoleCommand:
    properties:
      description:
        description: 描述
        type: string
      id:
        description: 角色ID
        type: integer
      localize:
        description: 多语言标识
        type: string
      name:
        description: 角色名称
        type: string
      sequence:
        description: 排序
        type: integer
      status:
        description: 状态
        type: integer
    required:
    - id
    - name
    type: object
...
  dto.DataPermissionDto:
    properties:
      deptIds:
        description: 自定义部门ID列表
        items:
          type: string
        type: array
      id:
        description: ID
        type: string
      roleId:
        description: 角色ID
        type: integer
      scope:
        description: 数据范围(1:全部数据 2:本部门数据 3:本部门及下级数据 4:仅本人数据 5:自定义部门数据)
        type: integer
      tenantId:
        description: 租户ID
        type: string
    type: object
...
  dto.RoleDto:
    properties:
      code:
        description: 角色代码
        type: string
      createdAt:
        description: 创建时间
        type: integer
      description:
        description: 描述
        type: string
      id:
        description: 角色ID
        type: integer
      localize:
        description: 国际化key
        type: string
      name:
        description: 角色名称
        type: string
      permIds:
        description: 权限id
        items:
          type: integer
        type: array
      sequence:
        description: 排序
        type: integer
      status:
        description: 状态
        type: integer
      tenantId:
        description: 租户ID
        type: string
      type:
        description: 角色类型(1:资源角色 2:数据权限角色)
        type: integer
      updatedAt:
        description: 更新时间
        type: integer
    type: object
...
host: localhost:8888
info:
  contact:
    name: go-server-template-admin
  description: This is a demo using go-server-template-admin.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: ares-ddd-admin
  version: "1.0"
paths:
...
  /v1/data-permission/{id}:
    get:
      consumes:
      - application/json
      description: 获取指定角色ID的数据权限配置
      parameters:
      - description: 角色ID
        in: path
        name: roleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/base_info.Success'
            - properties:
                data:
                  $ref: '#/definitions/dto.DataPermissionDto'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 获取角色的数据权限
      tags:
      - 数据权限
...
  /v1/sys/role:
    get:
      description: 获取角色
      operationId: RoleList
      parameters:
      - description: 编码
        in: query
        name: code
        type: string
      - description: 页码，从1开始
        in: query
        name: current
        type: integer
      - description: 名称
        in: query
        name: name
        type: string
      - description: 页码大小，最大100
        in: query
        name: size
        type: integer
      - description: 角色状态（禁用、启用）
        in: query
        name: status
        type: integer
      - description: 类型
        in: query
        name: type
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/base_info.Success'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.RoleDto'
                  type: array
              type: object
        "400":
          description: code为400 参数输入错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: code为401 token未带上
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: code为500 服务端内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 获取角色
      tags:
      - 系统角色
    post:
      description: 添加角色
      operationId: AddRole
      parameters:
      - description: 属性说明请在对应model中查看
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/commands.CreateRoleCommand'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/base_info.Success'
        "400":
          description: code为400 参数输入错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: code为401 token未带上
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: code为500 服务端内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 添加角色
      tags:
      - 系统角色
    put:
      consumes:
      - application/json
      description: 更新角色信息，包括基本信息和权限关联
      operationId: UpdateRole
      parameters:
      - description: 角色更新信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/commands.UpdateRoleCommand'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/base_info.Success'
        "400":
          description: code为400 参数输入错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: code为401 token未带上
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: code为500 服务端内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 更新角色
      tags:
      - 系统角色
  /v1/sys/role/:id:
    get:
      consumes:
      - application/json
      description: 根据id获取角色
      operationId: GetRoleDetails
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/base_info.Success'
            - properties:
                data:
                  $ref: '#/definitions/dto.RoleDto'
              type: object
        "400":
          description: code为400 参数输入错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: code为401 token未带上
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: code为500 服务端内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 根据id获取角色
      tags:
      - 系统角色
  /v1/sys/role/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定ID的角色
      operationId: DeleteRole
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/base_info.Success'
        "400":
          description: code为400 参数输入错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: code为401 token未带上
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: code为500 服务端内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 删除角色
      tags:
      - 系统角色
  /v1/sys/role/data-permission:
    get:
      consumes:
      - application/json
      description: 获取系统中所有数据权限类型的角色列表
      operationId: GetAllDataPermissionRoles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/base_info.Success'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.RoleDto'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 获取所有数据权限角色
      tags:
      - 系统角色
  /v1/sys/role/enabled:
    get:
      consumes:
      - application/json
      description: 获取系统中所有启用状态的角色列表
      operationId: GetAllEnabledRoles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/base_info.Success'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.RoleDto'
                  type: array
              type: object
        "400":
          description: code为400 参数输入错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: code为401 token未带上
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: code为500 服务端内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 获取所有启用状态的角色
      tags:
      - 系统角色
...
schemes:
- http
swagger: "2.0"